+++ CHANGELOG.md
@@ -20,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
-TBD
+- Use of external EEPROM
 
 ## [2.1.0] - 2025.11.30